requirements.txt
numpy>=1.22
scikit-learn>=1.3
faiss-cpu>=1.7.4
PyPDF2>=3.0
reportlab>=3.6
pytest>=7.0